minor tweaks for aix build

+++ src/common/daemonize.c
@@ -211,7 +211,7 @@ test_core_limit(void)
 		rlim->rlim_cur /= 1024;	/* bytes to KB */
 		if (rlim->rlim_cur < 2048) {
 			verbose("Warning: Core limit is only %ld KB",
-				rlim->rlim_cur);
+				(long int) rlim->rlim_cur);
 		}
 	}
 #endif
